+++ Source/Viewer/source/HexModeLayout.cpp
@@ -307,7 +307,7 @@ std::pair<double, double> HexModeLayout::CalcColumnSelectionBackground(const CFR
 
     const auto symb_width = m_Frame->FontInfo().PreciseMonospaceWidth();
     auto x1 = _offsets.columns.at(_columm_index) + local_start_byte * symb_width * 3;
-    auto x2 = x1 + sel_range.length * symb_width * 3 - symb_width;
+    auto x2 = x1 + static_cast<double>(sel_range.length) * symb_width * 3 - symb_width;
     return {std::floor(x1), std::ceil(x2)};
 }
